requirements.txt
pydantic
lark-parser
openai
backoff